added resourcepicker in common
