+++ libs/designer/src/lib/ui/common/resourcepicker/resourcepicker.tsx
@@ -1,33 +1,43 @@
-import type { AppDispatch, RootState } from '../../../core/state/templates/store';
-import { useDispatch, useSelector } from 'react-redux';
 import { useIntl } from 'react-intl';
-import { Option, Field, Dropdown } from '@fluentui/react-components';
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useCallback, useMemo } from 'react';
 import { useLocations, useLogicApps, useResourceGroups, useSubscriptions } from '../../../core/queries/resource';
-import {
-  setLocation,
-  setLogicAppDetails,
-  setResourceGroup,
-  setSubscription,
-  setWorkflowAppDetails,
-} from '../../../core/state/templates/workflowSlice';
-import { type LogicAppResource, type Resource, type Template, equals } from '@microsoft/logic-apps-shared';
-import { useTemplatesStrings } from '../templatesStrings';
+import { type LogicAppResource, type Template, equals } from '@microsoft/logic-apps-shared';
 import { useAllLogicApps } from '../../../core/configuretemplate/utils/queries';
-// import { ResourcePicker } from '../../common/resourcepicker/resourcepicker';
+import { useTemplatesStrings } from './../../templates/templatesStrings';
+import { ResourceField } from './resourcefield';
 
 export interface ResourcePickerProps {
   viewMode?: 'default' | 'alllogicapps';
   onSelectApp?: (value: LogicAppResource) => void;
   lockField?: Template.ResourceFieldId;
+  resourceState: {
+    subscriptionId: string;
+    resourceGroup: string;
+    location: string;
+    workflowAppName: string | undefined;
+    logicAppName: string | undefined;
+    isConsumption: boolean | undefined;
+  };
+  onSubscriptionSelect: (value: string) => void;
+  onResourceGroupSelect: (value: string) => void;
+  onLocationSelect: (value: string) => void;
+  onLogicAppSelect: (value: { name: string; location: string }) => void;
+  onLogicAppInstanceSelect: (value: { name: string; location: string; plan: string }) => void;
 }
 
-export const TemplateResourcePicker = ({ viewMode = 'default', onSelectApp, lockField }: ResourcePickerProps) => {
-  const dispatch = useDispatch<AppDispatch>();
+export const ResourcePicker = ({
+  viewMode = 'default',
+  onSelectApp,
+  lockField,
+  resourceState,
+  onSubscriptionSelect,
+  onResourceGroupSelect,
+  onLocationSelect,
+  onLogicAppSelect,
+  onLogicAppInstanceSelect,
+}: ResourcePickerProps) => {
   const isDefaultMode = viewMode === 'default';
-  const { subscriptionId, resourceGroup, location, workflowAppName, logicAppName, isConsumption } = useSelector(
-    (state: RootState) => state.workflow
-  );
+  const { subscriptionId, resourceGroup, location, workflowAppName, logicAppName, isConsumption } = resourceState;
   const { data: subscriptions, isLoading } = useSubscriptions();
   const { data: resourceGroups, isLoading: isResourceGroupLoading } = useResourceGroups(subscriptionId ?? '');
   const { data: locations, isLoading: islocationLoading } = useLocations(subscriptionId ?? '');
@@ -65,35 +75,32 @@ export const TemplateResourcePicker = ({ viewMode = 'default', onSelectApp, lock
   );
 
   const { resourceStrings } = useTemplatesStrings();
-  const onLogicAppSelect = useCallback(
+  const handleLogicAppSelect = useCallback(
     (value: string) => {
       const app = logicApps?.find((app) => equals(app.name, value));
-      dispatch(setWorkflowAppDetails({ name: value, location: app?.location ?? '' }));
+      onLogicAppSelect({ name: value, location: app?.location ?? '' });
     },
-    [dispatch, logicApps]
+    [onLogicAppSelect, logicApps]
   );
 
-  const onLogicAppInstanceSelect = useCallback(
+  const handleLogicAppInstanceSelect = useCallback(
     (value: string) => {
       const app = allLogicApps?.find((app) => equals(app.name, value));
-      dispatch(setLogicAppDetails({ name: value, location: app?.location ?? '', plan: app?.plan ?? '' }));
+      onLogicAppInstanceSelect({ name: value, location: app?.location ?? '', plan: app?.plan ?? '' });
 
       if (app) {
         onSelectApp?.(app);
       }
     },
-    [dispatch, allLogicApps, onSelectApp]
+    [onLogicAppInstanceSelect, allLogicApps, onSelectApp]
   );
 
   return (
     <div>
-      {/* <ResourcePicker
-        {...props}
-      /> */}
       <ResourceField
         id="subscriptionId"
         label={resourceStrings.SUBSCRIPTION}
-        onSelect={(value) => dispatch(setSubscription(value))}
+        onSelect={onSubscriptionSelect}
         defaultKey={subscriptionId}
         isLoading={isLoading}
         resources={subscriptions ?? []}
@@ -103,7 +110,7 @@ export const TemplateResourcePicker = ({ viewMode = 'default', onSelectApp, lock
       <ResourceField
         id="resourceGroupName"
         label={resourceStrings.RESOURCE_GROUP}
-        onSelect={(value) => dispatch(setResourceGroup(value))}
+        onSelect={onResourceGroupSelect}
         defaultKey={resourceGroup}
         isLoading={isResourceGroupLoading}
         resources={resourceGroups ?? []}
@@ -114,7 +121,7 @@ export const TemplateResourcePicker = ({ viewMode = 'default', onSelectApp, lock
         <ResourceField
           id="location"
           label={resourceStrings.LOCATION}
-          onSelect={(value) => dispatch(setLocation(value))}
+          onSelect={onLocationSelect}
           defaultKey={location}
           isLoading={islocationLoading}
           resources={locations ?? []}
@@ -126,7 +133,7 @@ export const TemplateResourcePicker = ({ viewMode = 'default', onSelectApp, lock
         <ResourceField
           id="logicapp"
           label={resourceStrings.LOGIC_APP}
-          onSelect={onLogicAppSelect}
+          onSelect={handleLogicAppSelect}
           defaultKey={workflowAppName ?? ''}
           isLoading={isLogicAppsLoading}
           resources={(logicApps ?? []).map((app) => ({
@@ -142,7 +149,7 @@ export const TemplateResourcePicker = ({ viewMode = 'default', onSelectApp, lock
         <ResourceField
           id="alllogicapp"
           label={intlText.ALL_LOGIC_APPS}
-          onSelect={onLogicAppInstanceSelect}
+          onSelect={handleLogicAppInstanceSelect}
           defaultKey={logicAppName ?? ''}
           isLoading={isAllLogicAppsLoading}
           resources={(allLogicApps ?? []).map((app) => ({
@@ -154,91 +161,6 @@ export const TemplateResourcePicker = ({ viewMode = 'default', onSelectApp, lock
           lockField={lockField === 'resource'}
         />
       )}
-    </div>
-  );
-};
-
-const ResourceField = ({
-  id,
-  label,
-  resources,
-  defaultKey,
-  errorMessage,
-  isLoading,
-  onSelect,
-  lockField,
-}: {
-  id: string;
-  label: string;
-  defaultKey: string;
-  resources: Resource[];
-  onSelect: (value: any) => void;
-  isLoading?: boolean;
-  errorMessage?: string;
-  lockField: boolean;
-}) => {
-  const intl = useIntl();
-  const texts = {
-    LOADING: intl.formatMessage({
-      defaultMessage: 'Loading resources ...',
-      id: 'IMWSjN',
-      description: 'Loading text',
-    }),
-    NO_ITEMS: intl.formatMessage({
-      defaultMessage: 'No resources found',
-      id: 'yytPY3',
-      description: 'No items to select text',
-    }),
-  };
-
-  const sortedResources = useMemo(() => resources.sort((a, b) => a.displayName.localeCompare(b.displayName)), [resources]);
-
-  const [selectedResource, setSelectedResource] = useState<string | undefined>('');
-  useEffect(() => {
-    if (!isLoading) {
-      const resource = resources.find((resource) => equals(resource.name, defaultKey))?.displayName;
-      if (!resource && !!defaultKey) {
-        onSelect('');
-      }
-
-      if (resource !== selectedResource) {
-        setSelectedResource(resource);
-      }
-    }
-  }, [resources, defaultKey, onSelect, isLoading, selectedResource]);
-
-  return (
-    <div style={{ marginBottom: '12px' }}>
-      <Field
-        className="msla-templates-tab-label"
-        label={label}
-        required={true}
-        validationMessage={errorMessage}
-        validationState={errorMessage ? 'error' : 'none'}
-      >
-        <Dropdown
-          style={{ width: '100%' }}
-          id={id}
-          onOptionSelect={(e, option) => onSelect(option?.optionValue)}
-          disabled={isLoading || (lockField && !!selectedResource)}
-          value={selectedResource}
-          selectedOptions={[defaultKey]}
-          size="small"
-          placeholder={isLoading ? texts.LOADING : ''}
-        >
-          {!isLoading && !sortedResources.length ? (
-            <Option key={'no-items'} value={'#noitem#'} disabled>
-              {texts.NO_ITEMS}
-            </Option>
-          ) : (
-            sortedResources.map((resource) => (
-              <Option key={resource.id} value={resource.name}>
-                {resource.displayName}
-              </Option>
-            ))
-          )}
-        </Dropdown>
-      </Field>
     </div>
   );
 };
